+++ server_utils.cpp
@@ -2,7 +2,7 @@
 
 #include "error.h"
 
-void emit_buffer_to_file(char *file, unsigned char *data, int n_bytes)
+void emit_buffer_to_file(char *file, unsigned char *data, size_t n_bytes)
 {
 	FILE *fh = fopen(file, "a+");
 	if (!fh)
